Update README.md
I have extended the rule regarding significant digits. There was a problem with the original rule.
Let's look at two examples: 10.5±0.1 and 10.5±0.5.
In the first example, we know that the last digit can vary by 1, so it could be 4, 5 or 6. We know that the last digit is definitely not 1 and definitely not 9. So while there is some uncertainty, it's not completely random.
In the second example, the last digit could be anything from 0 to 9. If it could be anything, that means it's just noise.