Update significance_table.py
Removed "Number of tests", it was inaccurate and I don't want to bother fixing it. P-values for n<50 are now calculated exactly rather than approximately. They are also formatted based on their value. If p-value is >0.01, the color is grey, and the conclusion is that we cannot tell which algorithm performs better.